Django_Internship/requirements.txt
--index-url https://download.pytorch.org/whl/cu124
--extra-index-url https://pypi.org/simple
Django
torch
torchvision
Pillow
